+++ Payroll/Transaction/AddCommissiondEmployee.cs
@@ -6,10 +6,10 @@ namespace Payroll.Test
 {
 	class AddCommissiondEmployee : AddEmployeeTransaction
 	{
-		private readonly float percent;
+		private readonly double percent;
 		private readonly double salary;
 
-		public AddCommissiondEmployee (int id, string name, string address, double salary, float percent) : base (id, name, address)
+		public AddCommissiondEmployee (int id, string name, string address, double salary, double percent) : base (id, name, address)
 		{
 			this.salary = salary;
 			this.percent = percent;
